Move useColumns to funding helper file
