+++ moped-editor/src/views/projects/projectView/ProjectFunding/helpers.js
@@ -1,3 +1,16 @@
+import { useMemo } from "react";
+import { Typography, Stack, IconButton } from "@mui/material";
+import LookupAutocompleteComponent from "src/components/DataGridPro/LookupAutocompleteComponent";
+import DataGridTextField from "src/components/DataGridPro/DataGridTextField";
+import ViewOnlyTextField from "src/components/DataGridPro/ViewOnlyTextField";
+import DollarAmountIntegerField from "src/views/projects/projectView/ProjectFunding/DollarAmountIntegerField";
+import EditOutlinedIcon from "@mui/icons-material/EditOutlined";
+import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
+import AttachFileOutlinedIcon from "@mui/icons-material/AttachFileOutlined";
+import ProjectFileLink from "src/views/projects/projectView/ProjectFiles/ProjectFileLink";
+import DataGridActions from "src/components/DataGridPro/DataGridActions";
+import { currencyFormatter } from "src/utils/numberFormatters";
+
 /** Transforms database funding records to DataGrid rows with lookup objects to populate autocomplete components
  * @param {Array} fundingRecords - array of funding records from the database
  * @param {Object} lookupData - object containing lookup arrays from the database
@@ -126,3 +139,267 @@ export const transformGridToDatabase = (gridRecord) => {
     funding_amount,
   };
 };
+
+// object to pass to the Fund column's LookupAutocomplete component
+const fduAutocompleteProps = {
+  getOptionLabel: (option) =>
+    option.fdu ? `${option.fdu} - ${option.unit_long_name}` : "",
+  isOptionEqualToValue: (value, option) =>
+    value?.ecapris_funding_id === option?.ecapris_funding_id,
+};
+
+const fduAutocompleteDependentFields = [
+  {
+    fieldName: "unit_long_name",
+    setFieldValue: (newValue) => newValue?.unit_long_name,
+  },
+  {
+    fieldName: "fund_source",
+    setFieldValue: (newValue) => newValue?.moped_fund_source,
+  },
+  {
+    fieldName: "fund_program",
+    setFieldValue: (newValue) => newValue?.moped_fund_program,
+  },
+  {
+    fieldName: "funding_amount",
+    setFieldValue: (newValue) => newValue?.amount,
+  },
+];
+
+export const isCellEditable = (params) => {
+  if (params.row.is_synced_from_ecapris) {
+    return false;
+  } else {
+    // if record is not synced from ecapris, but is also not manual, it means its been overriden
+    // dont edit using data grid
+    if (!params.row.is_manual) {
+      return false;
+    }
+    // records that are not synced from ecapris and are manual are editable
+    return true;
+  }
+};
+
+/** Hook that provides memoized column settings */
+export const useColumns = ({
+  dataProjectFunding,
+  dataLookups,
+  rowModesModel,
+  handleDeleteOpen,
+  handleSaveClick,
+  handleCancelClick,
+  handleEditClick,
+  handleAttachmentClick,
+  setOverrideFundingRecord,
+  usingShiftKey,
+  logUserEvent,
+}) =>
+  useMemo(() => {
+    return [
+      {
+        headerName: "FDU",
+        field: "fdu",
+        width: 180,
+        editable: true,
+        renderCell: ({ row, value }) =>
+          row.is_synced_from_ecapris ? (
+            <>
+              <span>{value?.fdu}</span>
+              <Typography
+                variant="body2"
+                sx={{
+                  color: "primary.main",
+                  fontWeight: 500,
+                }}
+              >
+                SYNCED FROM ECAPRIS
+              </Typography>
+            </>
+          ) : (
+            value?.fdu
+          ),
+        renderEditCell: (props) => (
+          <LookupAutocompleteComponent
+            {...props}
+            name={"ecapris_funding"}
+            options={dataLookups?.ecapris_subproject_funding}
+            fullWidthPopper={true}
+            autocompleteProps={{
+              ...fduAutocompleteProps,
+              value: props?.row?.fdu,
+            }}
+            dependentFieldsArray={fduAutocompleteDependentFields}
+          />
+        ),
+      },
+      {
+        headerName: "Unit Name",
+        field: "unit_long_name",
+        editable: true, // this is to be able to use the renderEditCell option to update the related phase
+        // during editing -- the input field is always disabled
+        width: 175,
+        renderEditCell: (props) => (
+          <ViewOnlyTextField
+            {...props}
+            value={props.row.unit_long_name}
+            usingShiftKey={usingShiftKey}
+            previousColumnField="fdu"
+            nextColumnField="amount"
+          />
+        ),
+      },
+      {
+        headerName: "Source",
+        field: "fund_source",
+        width: 180,
+        editable: true,
+        valueFormatter: (value) => value?.funding_source_name,
+        renderEditCell: (props) => (
+          <LookupAutocompleteComponent
+            {...props}
+            name={"funding_source"}
+            options={dataProjectFunding?.moped_fund_sources ?? []}
+            fullWidthPopper={true}
+          />
+        ),
+      },
+      {
+        headerName: "Program",
+        field: "fund_program",
+        width: 180,
+        editable: true,
+        valueFormatter: (value) => value?.funding_program_name,
+        renderEditCell: (props) => (
+          <LookupAutocompleteComponent
+            {...props}
+            name={"funding_program"}
+            options={dataProjectFunding?.moped_fund_programs ?? []}
+            fullWidthPopper={true}
+          />
+        ),
+      },
+      {
+        headerName: "Description",
+        field: "funding_description",
+        width: 200,
+        editable: true,
+        renderEditCell: (props) => <DataGridTextField {...props} multiline />,
+      },
+      {
+        headerName: "Status",
+        field: "fund_status",
+        editable: true,
+        width: 100,
+        valueFormatter: (value) => value?.funding_status_name,
+        renderEditCell: (props) => (
+          <LookupAutocompleteComponent
+            {...props}
+            name={"funding_status"}
+            defaultValue={1}
+            options={dataProjectFunding?.moped_fund_status ?? []}
+            fullWidthPopper={true}
+          />
+        ),
+      },
+      {
+        headerName: "Amount",
+        field: "funding_amount",
+        width: 100,
+        editable: true,
+        valueFormatter: (value) =>
+          value === null ? null : currencyFormatter.format(value),
+        renderEditCell: (props) => <DollarAmountIntegerField {...props} />,
+        type: "currency",
+      },
+      {
+        headerName: "Files",
+        field: "file_url",
+        minWidth: 150,
+        flex: 1,
+        editable: false,
+        renderCell: ({ row }) => {
+          return (
+            <Stack direction="column" spacing={0.5}>
+              {row.ecapris_funding_files.map((file_record) => {
+                const file = file_record.moped_project_file;
+
+                if (!file) return null;
+                return (
+                  <ProjectFileLink
+                    key={file.project_file_id}
+                    file_key={file.file_key}
+                    file_url={file.file_url}
+                  />
+                );
+              })}
+            </Stack>
+          );
+        },
+      },
+      {
+        headerName: "",
+        field: "edit",
+        hideable: false,
+        filterable: false,
+        sortable: false,
+        editable: false,
+        width: 110,
+        type: "actions",
+        renderCell: ({ id, row }) =>
+          row.is_manual ? (
+            <DataGridActions
+              id={id}
+              rowModesModel={rowModesModel}
+              handleCancelClick={handleCancelClick}
+              handleDeleteOpen={handleDeleteOpen}
+              handleSaveClick={handleSaveClick}
+              handleEditClick={handleEditClick}
+              editDisabled={row.is_synced_from_ecapris}
+              deleteDisabled={row.is_synced_from_ecapris}
+              handleAttachmentClick={handleAttachmentClick}
+            />
+          ) : (
+            <>
+              <IconButton
+                aria-label="edit"
+                sx={{ color: "inherit", padding: "5px" }}
+                onClick={() => {
+                  logUserEvent("funding_ecapris_override_form_load");
+                  setOverrideFundingRecord(row);
+                }}
+              >
+                <EditOutlinedIcon />
+              </IconButton>
+              <IconButton
+                aria-label="attachment"
+                sx={{ color: "inherit", padding: "5px" }}
+                onClick={handleAttachmentClick(id)}
+              >
+                <AttachFileOutlinedIcon />
+              </IconButton>
+              <IconButton
+                aria-label="delete"
+                sx={{ color: "inherit", padding: "5px" }}
+                disabled={!!row.is_synced_from_ecapris}
+                onClick={handleDeleteOpen(id)}
+              >
+                <DeleteOutlineIcon />
+              </IconButton>
+            </>
+          ),
+      },
+    ];
+  }, [
+    dataProjectFunding,
+    dataLookups,
+    rowModesModel,
+    handleDeleteOpen,
+    handleSaveClick,
+    handleCancelClick,
+    handleEditClick,
+    handleAttachmentClick,
+    setOverrideFundingRecord,
+    usingShiftKey,
+    logUserEvent,
+  ]);
